+++ nbReman/src/org/harley/reman/interfaz/interfaces/VMEduccion.java
@@ -141,6 +141,7 @@ public class VMEduccion extends JDialog {
         jTable3 = new javax.swing.JTable();
         btnVEDCancelar = new javax.swing.JButton();
         btnVEDGuardar = new javax.swing.JButton();
+        btnVEDGuardar1 = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
         setTitle("Educción");
@@ -206,6 +207,8 @@ public class VMEduccion extends JDialog {
 
         jLabel15.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
         jLabel15.setText("Nº");
+
+        txtEDVersion.setEditable(false);
 
         jLabel9.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
         jLabel9.setText("Fecha");
@@ -588,6 +591,14 @@ public class VMEduccion extends JDialog {
             }
         });
 
+        btnVEDGuardar1.setFont(new java.awt.Font("Tahoma", 1, 12)); // NOI18N
+        btnVEDGuardar1.setText("Versionar");
+        btnVEDGuardar1.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnVEDGuardar1ActionPerformed(evt);
+            }
+        });
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
@@ -595,9 +606,11 @@ public class VMEduccion extends JDialog {
             .addGroup(layout.createSequentialGroup()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(layout.createSequentialGroup()
-                        .addGap(250, 250, 250)
+                        .addGap(162, 162, 162)
+                        .addComponent(btnVEDGuardar1)
+                        .addGap(18, 18, 18)
                         .addComponent(btnVEDGuardar)
-                        .addGap(29, 29, 29)
+                        .addGap(18, 18, 18)
                         .addComponent(btnVEDCancelar))
                     .addGroup(layout.createSequentialGroup()
                         .addContainerGap()
@@ -610,9 +623,10 @@ public class VMEduccion extends JDialog {
                 .addContainerGap()
                 .addComponent(jTabbedPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(9, 9, 9)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(btnVEDCancelar)
                     .addComponent(btnVEDGuardar)
-                    .addComponent(btnVEDCancelar))
+                    .addComponent(btnVEDGuardar1))
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
@@ -645,10 +659,6 @@ public class VMEduccion extends JDialog {
                 || !ToolsInterface.verificarVersion(eduVer)) {
             error = true;
         }
-        if (!ToolsSystem.CompararVersiones(eduVer, myEdu.getEduVer())) {
-            ToolsInterface.msjError(this, "La version debe ser mayor a la anterior!");
-            return;
-        }
 
         if (!error) {
             if (sysReman.modificarEduccion(eduCod, eduNom, eduVer, eduTip, eduObj, eduFec, eduFueNom, eduFueCar, eduFueTip, eduEspNom, eduEspEsp, eduEspExp, eduEspCar, eduDes, eduObs)) {
@@ -678,9 +688,14 @@ public class VMEduccion extends JDialog {
 
     }//GEN-LAST:event_cmbEDFuenteActionPerformed
 
+    private void btnVEDGuardar1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnVEDGuardar1ActionPerformed
+        
+    }//GEN-LAST:event_btnVEDGuardar1ActionPerformed
+
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton btnVEDCancelar;
     private javax.swing.JButton btnVEDGuardar;
+    private javax.swing.JButton btnVEDGuardar1;
     private javax.swing.ButtonGroup buttonGroup1;
     private javax.swing.JComboBox cmbEDEspecialista;
     private javax.swing.JComboBox cmbEDFuente;
