+++ nbReman/src/org/harley/reman/interfaz/interfaces/VOExportar.form
@@ -36,11 +36,11 @@
                   <Component id="jLabel2" min="-2" max="-2" attributes="0"/>
                   <Component id="jLabel1" min="-2" max="-2" attributes="0"/>
               </Group>
-              <EmptySpace min="-2" pref="35" max="-2" attributes="0"/>
+              <EmptySpace min="-2" pref="26" max="-2" attributes="0"/>
               <Group type="103" groupAlignment="1" attributes="0">
                   <Group type="102" alignment="1" attributes="0">
                       <Component id="btnEExportar" min="-2" pref="98" max="-2" attributes="0"/>
-                      <EmptySpace type="separate" pref="30" max="32767" attributes="0"/>
+                      <EmptySpace type="separate" pref="39" max="32767" attributes="0"/>
                       <Component id="btnECancelar" min="-2" pref="97" max="-2" attributes="0"/>
                   </Group>
                   <Group type="102" alignment="0" attributes="0">
